IL-287 renamed `intelligence_app.py` to `auth_service.py`
