+++ src/intelligence_layer/core/auth_service.py
